fix decimal

+++ database/migrations/2026_01_12_144122_create_orders_table.php
@@ -18,7 +18,7 @@ return new class extends Migration
                 ->constrained('users')
                 ->cascadeOnDelete();
 
-            $table->integer('amount');
+            $table->decimal('amount', 8, 2);
             $table->string('status');
 
             $table->timestamps();
